refactor: split out messages into separate package
