+++ contracts/collectxyz-nft-contract/examples/schema.rs
@@ -3,7 +3,7 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use collectxyz_nft_contract::msg::{
+use collectxyz::nft::{
     ExecuteMsg, InstantiateMsg, MoveParamsResponse, QueryMsg, XyzTokensResponse,
 };
