+++ contracts/collectxyz-nft-contract/schema/instantiate_msg.json
