+++ contracts/collectxyz-nft-contract/schema/move_params_response.json
