+++ contracts/collectxyz-nft-contract/src/error.rs
