+++ contracts/collectxyz-nft-contract/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod contract;
 pub mod error;
 pub mod execute;
-pub mod msg;
 pub mod query;
 pub mod state;
