+++ packages/collectxyz/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "collectxyz"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+schemars = "0.8.3"
+serde = { version = "1.0.127", default-features = false, features = ["derive"] }
+cosmwasm-std = { version = "0.16.0" }
+cw721 = { version = "=0.9.1" }
+cw721-base = { version = "=0.9.1", features = ["library"] }
